+++ src/main/java/org/osgl/mvc/annotation/ResponseStatus.java
@@ -20,10 +20,7 @@ package org.osgl.mvc.annotation;
  * #L%
  */
 
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
+import java.lang.annotation.*;
 
 /**
  * Mark a certain class so when the implementation framework encountered a
@@ -31,6 +28,7 @@ import java.lang.annotation.Target;
  */
 @Retention(RetentionPolicy.RUNTIME)
 @Target(ElementType.METHOD)
+@Inherited
 public @interface ResponseStatus {
     /**
      * Specify the response status code